Add more test cases for query url/parameter sanitization
