+++ src/tests/utils/sanitize-query-parameter-sanitization.spec.tsx
@@ -1,102 +1,6 @@
 import {
-  isAllAlpha, sanitizeQueryParameter
+  sanitizeQueryParameter
 } from '../../app/utils/query-parameter-sanitization';
-import {
-  isDeprecation, sanitizeQueryUrl
-} from '../../app/utils/query-url-sanitization';
-
-
-describe('isAllAlpha should ', () => {
-  const list = [
-    { key: 'aaa', isAllAlphabetic: true },
-    { key: 'aZa', isAllAlphabetic: true },
-    { key: '111', isAllAlphabetic: false },
-    { key: '1a1', isAllAlphabetic: false }
-  ];
-
-  list.forEach(element => {
-    it(`return ${element.isAllAlphabetic} for ${element.key}`, () => {
-      const key = isAllAlpha(element.key);
-      expect(key).toBe(element.isAllAlphabetic);
-    });
-  });
-});
-
-describe('isDepraction should ', () => {
-  const list = [
-    { key: 'messages_V2', deprecated: true },
-    { key: 'messages', deprecated: false },
-    { key: 'users_v2', deprecated: true },
-    { key: 'users', deprecated: false }
-  ];
-
-  list.forEach(element => {
-    it(`return ${element.deprecated} for ${element.key}`, () => {
-      const key = isDeprecation(element.key);
-      expect(key).toBe(element.deprecated);
-    });
-  });
-});
-
-describe('Sanitize Query Url should', () => {
-  const list = [
-    {
-      check: 'without task id',
-      url: 'https://graph.microsoft.com/v1.0/planner/tasks/oIx3zN98jEmVOM-4mUJzSGUANeje',
-      sanitized: 'https://graph.microsoft.com/v1.0/planner/tasks/{tasks-id}',
-    },
-    {
-      check: 'without email',
-      url: 'https://graph.microsoft.com/v1.0/users/MiriamG@M365x214355.onmicrosoft.com',
-      sanitized: 'https://graph.microsoft.com/v1.0/users/{users-id}',
-    },
-    {
-      check: 'without message id',
-      // tslint:disable-next-line: max-line-length
-      url: 'https://graph.microsoft.com/v1.0/me/messages/AAMkAGVmMDEzMTM4LTZmYWUtNDdkNC1hMDZiLTU1OGY5OTZhYmY4OABGAAAAAAAiQ8W967B7TKBjgx9rVEURBwAiIsqMbYjsT5e-T7KzowPTAAAAAAEMAAAiIsqMbYjsT5e-T7KzowPTAAMCzwJpAAA=',
-      sanitized: 'https://graph.microsoft.com/v1.0/me/messages/{messages-id}'
-    },
-    {
-      check: 'without extension id',
-      url: 'https://graph.microsoft.com/v1.0/me/extensions/com.contoso.roamingSettings',
-      sanitized: 'https://graph.microsoft.com/v1.0/me/extensions/{extensions-id}'
-    },
-    {
-      check: 'without plan id',
-      url: 'https://graph.microsoft.com/v1.0/planner/plans/CONGZUWfGUu4msTgNP66e2UAAySi',
-      sanitized: 'https://graph.microsoft.com/v1.0/planner/plans/{plans-id}'
-    },
-    {
-      check: 'without section id',
-      url: 'https://graph.microsoft.com/v1.0/me/onenote/sections/1f7ff346-c174-45e5-af38-294e51d9969a/pages',
-      sanitized: 'https://graph.microsoft.com/v1.0/me/onenote/sections/{sections-id}/pages'
-    },
-    {
-      check: 'with all parameters replaced',
-      // tslint:disable-next-line: max-line-length
-      url: 'https://graph.microsoft.com/v1.0/teams/02bd9fd6-8f93-4758-87c3-1fb73740a315/channels/19:09fc54a3141a45d0bc769cf506d2e079@thread.skype',
-      sanitized: 'https://graph.microsoft.com/v1.0/teams/{teams-id}/channels/{channels-id}'
-    },
-    {
-      check: 'with deprecated resource and without id',
-      url: 'https://graph.microsoft.com/v1.0/applications_v2/02bd9fd6-8f93-4758-87c3-1fb73740a315',
-      sanitized: 'https://graph.microsoft.com/v1.0/applications_v2/{applications_v2-id}'
-    },
-    {
-      check: 'with id within brackets',
-      url: 'https://graph.microsoft.com/v1.0/users(\'48d31887-5fad-4d73-a9f5-3c356e68a038\')',
-      sanitized: 'https://graph.microsoft.com/v1.0/users(users-id)'
-    }
-  ];
-
-  list.forEach(element => {
-    it(`return url ${element.check}`, () => {
-      const sanitizedUrl = sanitizeQueryUrl(element.url);
-      expect(sanitizedUrl).toEqual(element.sanitized);
-    });
-  });
-
-});
 
 describe('Sanitize Query Parameters should', () => {
   const list = [
@@ -244,6 +148,37 @@ describe('Sanitize Query Parameters should', () => {
       queryParam: '$filter=(price sub 5) gt 10',
       sanitizedQueryParam: '$filter=(price sub <value>) gt <value>'
     },
+    {
+      check: 'returns sanitized html decoded value for $filter query option',
+      // tslint:disable-next-line: max-line-length
+      queryParam: '$filter=(creationDateTime%20ge%202017-06-25T07:00:00Z)%20and%20(creationDateTime%20le%202017-07-25T17:30:17Z)',
+      sanitizedQueryParam: '$filter=(creationDateTime ge <value>) and (creationDateTime le <value>)'
+    },
+    {
+      check: 'returns sanitized value with `any` collection operator for $filter query option',
+      queryParam: '$filter=identities/any(c:c/issuerAssignedId eq \'j.smith@yahoo.com\'',
+      sanitizedQueryParam: '$filter=identities/any(c:c/issuerAssignedId eq <value>'
+    },
+    {
+      check: 'returns sanitized value for $filter with `any` collection operator with fully qualified entity name',
+      queryParam: '$filter=microsoft.graph.countryNamedLocation/countriesAndRegions/any(c:%20c%20eq%20\'CA\')',
+      sanitizedQueryParam: '$filter=microsoft.graph.countryNamedLocation/countriesAndRegions/any(c: c eq <value>)'
+    },
+    {
+      check: 'returns sanitized value with `any` collection operator with no arguments for $filter query option',
+      queryParam: '$filter=not rooms/any()',
+      sanitizedQueryParam: '$filter=not rooms/any()'
+    },
+    {
+      check: 'returns sanitized value with `all` collection operator for $filter query option',
+      queryParam: '$filter=ratings/all(r: r ge 3 and r le 5)',
+      sanitizedQueryParam: '$filter=ratings/all(r: r ge <value> and r le <value>)'
+    },
+    {
+      check: 'returns sanitized value with complex `all` collection operator for $filter query option',
+      queryParam: '$filter=rooms/all(room: room/amenities/any(a: a eq \'tv\') and room/baseRate lt 100.0)',
+      sanitizedQueryParam: '$filter=rooms/all(room: room/amenities/any(a: a eq <value>) and room/baseRate lt <value>)'
+    },
 
     // $expand
     {
@@ -258,8 +193,13 @@ describe('Sanitize Query Parameters should', () => {
     },
     {
       check: 'returns property sanitized values for $expand value with multiple navigation properties',
-      queryParam: '$expand=children($select=firstname),customers($top="five")',
-      sanitizedQueryParam: '$expand=children($select=firstname),customers($top=<invalid-value>)'
+      queryParam: '$expand=children($select=firstname, lastname),customers($skip=10, $top="five")',
+      sanitizedQueryParam: '$expand=children($select=firstname),customers($skip=10, $top=<invalid-value>)'
+    },
+    {
+      check: 'returns value as is for fully qualified entity name',
+      queryParam: '$expand=microsoft.graph.itemattachment/item',
+      sanitizedQueryParam: '$expand=microsoft.graph.itemattachment/item'
     },
 
     // regular query parameters
@@ -273,6 +213,11 @@ describe('Sanitize Query Parameters should', () => {
       queryParam: 'someone@onmicrosoft.com=mail',
       sanitizedQueryParam: '<invalid-key>=<value>'
     },
+    {
+      check: 'returns value as is when query parameter key starts with $',
+      queryParam: '$level=max',
+      sanitizedQueryParam: '$level=<value>'
+    }
 
   ];
